# Docs/02_contracts_and_schemas/offisim_openapi.yaml
openapi: 3.1.0
info:
  title: Offisim Platform and Local Install APIs
  version: 0.2.0
  summary: Platform marketplace APIs plus local runtime install/import contract snapshots.
  description: >
    This file tracks the current public platform HTTP surface mounted from `apps/platform/src/index.ts`
    together with the local install transaction contract used by `packages/install-core`.
    Public marketplace reads are unauthenticated. Creator publishing, reviews, creator bootstrap,
    install receipt recording, and user library access require authenticated user context.
    Better Auth native session endpoints under `/api/auth/*` and API token management under
    `/v1/auth/tokens` are implementation-owned and intentionally not expanded here.
    Local install/import endpoints remain declaration-driven: no arbitrary install hooks,
    no postinstall scripts, and no remote code execution during import.
servers:
  - url: https://api.offisim.market
    description: Registry and marketplace API
  - url: http://127.0.0.1:43111
    description: Local desktop runtime API
...
security: []
paths:
...
  /v1/market/search:
    get:
      tags: [Market Public]
      summary: Search public marketplace listings
      operationId: searchMarketplace
      security: []
      parameters:
        - in: query
          name: q
          schema: { type: string, maxLength: 200 }
          description: Full-text query over title, summary, creator, and tags.
        - in: query
          name: kind
          schema:
            type: string
            enum: [employee, skill, sop, company_template, office_layout, bundle, prefab]
        - in: query
          name: risk_class
          schema:
            type: string
            enum: [data_asset, logic_asset, privileged_asset]
        - in: query
          name: tag
          schema: { type: string, maxLength: 32 }
        - in: query
          name: sort
          schema:
            type: string
            enum: [relevance, newest, updated, rating, installs]
            default: relevance
        - in: query
          name: page
          schema: { type: integer, minimum: 1, default: 1 }
        - in: query
          name: per_page
          schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
      responses:
        '200':
          description: Search result page
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
  /v1/market/listings/{listingId}:
    get:
      tags: [Market Public]
      summary: Get public marketplace listing detail
      operationId: getListingDetail
      security: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/market/listings/by-slug/{slug}:
    get:
      tags: [Market Public]
      summary: Get public marketplace listing detail by slug
      operationId: getListingDetailBySlug
      security: []
      parameters:
        - $ref: '#/components/parameters/Slug'
      responses:
        '200':
          description: Listing detail
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListingDetail'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/market/listings/{listingId}/versions:
    get:
      tags: [Market Public]
      summary: Get visible versions for a listing
      operationId: listListingVersions
      security: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Versions for a listing
          content:
            application/json:
              schema:
                type: object
                required: [listing_id, versions]
                properties:
                  listing_id: { type: string }
                  versions:
                    type: array
                    items:
                      $ref: '#/components/schemas/VersionSummary'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/market/listings/{listingId}/reviews:
    get:
      tags: [Market Public, Reviews]
      summary: List visible reviews for a listing
      operationId: listListingReviews
      security: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Review list
          content:
            application/json:
              schema:
                type: object
                required: [listing_id, reviews]
                properties:
                  listing_id: { type: string }
                  reviews:
                    type: array
                    items:
                      $ref: '#/components/schemas/Review'
  /v1/market/listings/{listingId}/reports:
    post:
      tags: [Market Public]
      summary: Report a listing for moderation review
      operationId: reportListing
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingReportRequest'
      responses:
        '201':
          description: Report created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModerationFlag'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Only the listing creator can change listing status
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The current user already reported this listing
          content:
            application/json:
              schema: { $ref: '#/components/schemas/ErrorResponse' }
  /v1/market/listings/{listingId}/forks:
    get:
      tags: [Market Public]
      summary: List listings that forked from this listing
      operationId: listListingForks
      security: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Fork listings
          content:
            application/json:
              schema:
                type: object
                required: [forks]
                properties:
                  forks:
                    type: array
                    items:
                      $ref: '#/components/schemas/ForkSummary'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/market/listings/{listingId}/lineage:
    get:
      tags: [Market Public]
      summary: Get lineage chain for a listing
      operationId: getListingLineage
      security: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      responses:
        '200':
          description: Listing lineage
          content:
            application/json:
              schema:
                type: object
                required: [ancestors, descendants]
                properties:
                  ancestors:
                    type: array
                    items:
                      $ref: '#/components/schemas/LineageNode'
                  descendants:
                    type: array
                    items:
                      $ref: '#/components/schemas/LineageNode'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/market/listings/{listingId}/status:
    patch:
      tags: [Creator Publish]
      summary: Change listing visibility status for the owning creator
      operationId: patchListingStatus
      security:
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/ListingId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ListingStatusPatchRequest'
      responses:
        '200':
          description: Listing status updated
          content:
            application/json:
              schema:
                type: object
                required: [ok, listing_id, status]
                properties:
                  ok: { type: boolean }
                  listing_id: { type: string }
                  status:
                    type: string
                    enum: [listed, hidden, retired]
        '400':
          $ref: '#/components/responses/ValidationError'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: API token or Better Auth bearer session
  parameters:
    ListingId:
      name: listingId
      in: path
      required: true
      schema: { type: string }
    Slug:
      name: slug
      in: path
      required: true
      schema: { type: string }
...
  responses:
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
    ValidationError:
      description: Validation failed
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ErrorResponse' }
  schemas:
    SearchResponse:
      type: object
      required: [items, page, per_page, total]
      properties:
        items:
          type: array
          items: { $ref: '#/components/schemas/ListingSummary' }
        page: { type: integer, minimum: 1 }
        per_page: { type: integer, minimum: 1 }
        total: { type: integer, minimum: 0 }
    ListingSummary:
      type: object
      required:
        [listing_id, slug, kind, title, summary, creator, status, latest_version, rating, install_count]
      properties:
        listing_id: { type: string }
        slug: { type: string }
        kind:
          type: string
          enum: [employee, skill, sop, company_template, office_layout, bundle, prefab]
        title: { type: string }
        summary: { type: string }
        creator:
          $ref: '#/components/schemas/CreatorSummary'
        status:
          type: string
          enum: [listed, hidden, retired]
        latest_version: { type: string }
        rating: { type: number, format: float }
        install_count: { type: integer, minimum: 0 }
        tags:
          type: array
          items: { type: string }
        preview:
          $ref: '#/components/schemas/PreviewRef'
    ListingDetail:
      allOf:
        - $ref: '#/components/schemas/ListingSummary'
        - type: object
          required: [description, requirements, permissions]
          properties:
            description: { type: string }
            version:
              anyOf:
                - $ref: '#/components/schemas/VersionSummary'
                - type: 'null'
            requirements:
              $ref: '#/components/schemas/RequirementsSummary'
            permissions:
              $ref: '#/components/schemas/PermissionSummary'
            lineage:
              $ref: '#/components/schemas/LineageSummary'
            previews:
              type: array
              items: { $ref: '#/components/schemas/PreviewRef' }
    VersionSummary:
      type: object
      required: [package_id, version, runtime_range, schema_version, environments, risk_class]
      properties:
        package_id: { type: string }
        package_version_id: { type: string }
        version: { type: string }
        runtime_range: { type: string }
        schema_version: { type: string }
        environments:
          type: array
          items:
            type: string
            enum: [desktop, docker, web_limited]
        risk_class:
          type: string
          enum: [data_asset, logic_asset, privileged_asset]
        published_at:
          type: string
          format: date-time
        changelog: { type: string }
    CreatorSummary:
      type: object
      required: [creator_id, handle, display_name, verification_state]
      properties:
        creator_id: { type: string }
        handle: { type: string }
        display_name: { type: string }
        verification_state:
          type: string
          enum: [unverified, verified, trusted]
...
    RequirementsSummary:
      type: object
      properties:
        required_capabilities:
          type: array
          items: { type: string }
        required_mcps:
          type: array
          items: { type: string }
        recommended_models:
          type: array
          items:
            $ref: '#/components/schemas/RecommendedModel'
    PermissionSummary:
      type: object
      properties:
        risk_class:
          type: string
          enum: [data_asset, logic_asset, privileged_asset]
        declares_secrets: { type: boolean }
        filesystem_scope:
          type: string
          enum: [none, workspace, project, custom_path]
        network_scope:
          type: string
          enum: [none, limited, unrestricted]
    RecommendedModel:
      type: object
      required: [profile]
      properties:
        profile: { type: string }
        reason: { type: string }
        provider_hints:
          type: array
          items: { type: string }
    LineageSummary:
      type: object
      properties:
        origin_package_id: { type: string }
        forked_from_version: { type: string }
        derivative_of:
          type: array
          items: { type: string }
    LineageNode:
      type: object
      required: [listingId, title, slug, version]
      properties:
        listingId: { type: string }
        title: { type: string }
        slug: { type: string }
        version: { type: string }
    ForkSummary:
      type: object
      required: [listingId, title, slug, creatorHandle, version, forkedAt]
      properties:
        listingId: { type: string }
        title: { type: string }
        slug: { type: string }
        creatorHandle: { type: string }
        version: { type: string }
        forkedAt:
          type: string
          format: date-time
    PreviewRef:
      type: object
      properties:
        kind:
          type: string
          enum: [icon, image, video, readme]
        url:
          type: string
          format: uri
        alt: { type: string }
...
    ListingReportRequest:
      type: object
      required: [reason]
      properties:
        reason:
          type: string
          enum: [spam, malicious_code, copyright, misleading, other]
        details:
          type: string
          maxLength: 1000
    ModerationFlag:
      type: object
      required: [flag_id, target_type, target_id, reason, status, created_at]
      properties:
        flag_id: { type: string }
        target_type: { type: string }
        target_id: { type: string }
        reason: { type: string }
        status: { type: string }
        created_at:
          type: string
          format: date-time
...
    ListingStatusPatchRequest:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [listed, hidden, retired]
        reason:
          type: string
          maxLength: 500
...
    Review:
      type: object
      required: [review_id, listing_id, rating, moderation_state, created_at, updated_at]
      properties:
        review_id: { type: string }
        listing_id: { type: string }
        user_id: { type: string }
        rating:
          type: integer
          minimum: 1
          maximum: 5
        title: { type: string }
        body: { type: string }
        moderation_state:
          type: string
          enum: [visible, hidden, flagged]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code: { type: string }
            message: { type: string }
            details:
              type: object
              additionalProperties: true